+++ conftest.py
@@ -55,3 +55,11 @@ def CLIENT(monkeypatch_session):
 
     monkeypatch_session.setattr(client, "_make_request", _make_request_patch)
     return client
+
+@pytest.fixture()
+def dataset(CLIENT):
+    ds = CLIENT.create_dataset(TEST_DATASET_NAME)
+    ds.append(TEST_DATASET_ITEMS)
+    yield ds
+
+    CLIENT.delete_dataset(ds.id)
